+++ app/initializers/store.js
@@ -2,10 +2,10 @@ export default {
   name: 'fireplace:inject-store',
 
   initialize(container, application) {
-    application.inject('controller',   'store', 'store:main');
-    application.inject('route',        'store', 'store:main');
-    application.inject('data-adapter', 'store', 'store:main');
-    application.inject('collection',   'store', 'store:main');
-    application.inject('component',    'store', 'store:main');
+    application.inject('controller',   'store', 'service:store');
+    application.inject('route',        'store', 'service:store');
+    application.inject('data-adapter', 'store', 'service:store');
+    application.inject('collection',   'store', 'service:store');
+    application.inject('component',    'store', 'service:store');
   }
 };