+++ app/services/store.js
@@ -1,0 +1,2 @@
+import Store from 'fireplace/store';
+export default Store;