+++ app/stores/main.js
@@ -1,3 +1,0 @@
-import { Store } from 'fireplace';
-
-export default Store;